Deduplicate CUDA detection; clean up dev_plans_archive docs
Extract the nvidia-smi parser shared by setup.py and scripts/install.py
into a top-level _cuda_detect module so both call sites depend on one
implementation. Add it to py-modules so non-editable installs work.

Retitle currentPYnJAXplan.md to reflect what shipped (Python 3.13;
3.14 deferred because TF lacks cp314 wheels) and add a status banner
pointing to the deferral section.

Rewrite multi_gpu_edge_case.txt as single_cuda_stack_check.md — a
proper design note covering the dual-plugin problem, the two-layer
defense, FABRICPC_ALLOW_MULTIPLE_CUDA semantics, and the venv-recreate
recovery recipe. The old name was misleading (this is not a multi-GPU
issue) and the content was a raw transcript with box-drawing tables.
